gcp/requirements.txt
tensorflow==2.18
google-cloud-storage==1.16.1
Pillow==6.0.0
Flask